Replace %zd with %lu for windows portability
git-svn-id: https://www.dnssec-tools.org/svn/dnssec-tools/trunk@5883 c4029f9c-b713-0410-a448-fae9e31a314a

+++ dnssec-tools/validator/apps/libsres_test.c
@@ -180,10 +180,13 @@ query_async_test(int async, int burst_max, int inflight_max, int numq)
         ++sent;
         if ((rc >= 0) || (SR_NO_ANSWER == rc)) {
             ++answered;
-            printf("sent %zd %s, got %d bytes\n", count, names[count], len);
+            printf("sent %lu %s, got %lu bytes\n", 
+                   (unsigned long)count, names[count], 
+                   (unsigned long)len);
         }
         else {
-            printf("bad rc %d/%zd bytes from get(%s) @ count %d\n", rc, len,
+            printf("bad rc %d/%lu bytes from get(%s) @ count %d\n", 
+                   rc, (unsigned long)len,
                    names[count], count);
         }
     }
